+++ src/app/(journal)/index.tsx
@@ -1,39 +1,5 @@
-import { PlatformColor, Pressable, useColorScheme } from "react-native";
+import { Redirect } from "expo-router";
 
-import { Stack, useRouter } from "expo-router";
-import { SymbolView } from "expo-symbols";
-
-import { TemplateView } from "@/components/template/template-view";
-
-/**
- * ジャーナル画面
- */
-export default function TemplateScreen() {
-  const router = useRouter();
-  const colorScheme = useColorScheme();
-  const titleColor = colorScheme === "dark" ? "#ffffff" : "#000000";
-
-  return (
-    <>
-      <Stack.Screen
-        options={{
-          title: "Journal",
-          headerTitleStyle: {
-            color: titleColor,
-          },
-          headerShown: true,
-          headerLargeTitleEnabled: true,
-          headerBackButtonDisplayMode: "minimal",
-          headerRight: () => (
-            <Pressable
-              onPress={() => router.push("/(journal)/template-create")}
-            >
-              <SymbolView name="plus" tintColor={PlatformColor("label")} />
-            </Pressable>
-          ),
-        }}
-      />
-      <TemplateView />
-    </>
-  );
+export default function JournalIndex() {
+  return <Redirect href="/(journal)/template" />;
 }
