+++ src/app/(journal)/template/create.tsx
