+++ src/components/entry/entry-list-view.tsx
