+++ src/components/entry/entry-row.tsx
